Add VERSION env var and update SonarQube analysis command
The `environment` block in the `pipeline` section of `TestJenkins` has been updated to include a new environment variable `VERSION` set to "0.5.0".

In the `stage('SonarQube Analysis')` section, the `sh` command for starting the SonarQube analysis has been modified. The new command now includes the `sonar.projectVersion` parameter, which is set to the value of the `VERSION` environment variable.