Handle missing movies gracefully in repository methods
The changes in `GetAllCuevanaMoviesCommand.cs` add a conditional check to skip processing for movies that do not have a corresponding entry in the movie repository.

In `MovieRepository.cs`, the `GetIdByTmdbId` method is modified to return 0 if no movie is found with the given `TmdbId`, instead of throwing an exception. This is achieved by using `FirstOrDefaultAsync` instead of `FirstAsync`.